update user-avatar, move file to search folder
