+++ lib/features/search/components/live-preview-card.tsx
