+++ lib/features/search/components/live-preview.tsx
